+++ src/main/scala/edu/ie3/simona/sim/setup/ExtSimSetup.scala
@@ -6,11 +6,9 @@
 
 package edu.ie3.simona.sim.setup
 
-import edu.ie3.simona.api.data.connection.{
-  ExtEvDataConnection,
-  ExtInputDataConnection,
-  ExtPrimaryDataConnection,
-}
+import com.typesafe.config.Config
+import edu.ie3.datamodel.models.input.container.JointGridContainer
+import edu.ie3.simona.api.data.connection.*
 import edu.ie3.simona.api.ontology.DataMessageFromExt
 import edu.ie3.simona.api.ontology.simulation.ControlResponseMessageFromExt
 import edu.ie3.simona.api.simulation.{ExtSimAdapterData, ExtSimulation}
@@ -19,6 +17,8 @@ import edu.ie3.simona.exceptions.ServiceException
 import edu.ie3.simona.ontology.messages.{SchedulerMessage, ServiceMessage}
 import edu.ie3.simona.scheduler.ScheduleLock
 import edu.ie3.simona.service.ServiceStateData.InitializeServiceStateData
+import edu.ie3.simona.service.em.ExtEmDataService
+import edu.ie3.simona.service.em.ExtEmDataService.InitExtEmData
 import edu.ie3.simona.service.ev.ExtEvDataService
 import edu.ie3.simona.service.ev.ExtEvDataService.InitExtEvData
 import edu.ie3.simona.util.SimonaConstants.PRE_INIT_TICK
@@ -26,6 +26,7 @@ import org.apache.pekko.actor.typed.ActorRef
 import org.apache.pekko.actor.typed.scaladsl.ActorContext
 import org.slf4j.{Logger, LoggerFactory}
 
+import java.time.ZonedDateTime
 import java.util.UUID
 import scala.jdk.CollectionConverters.{ListHasAsScala, SetHasAsScala}
 import scala.util.{Failure, Success, Try}
@@ -44,6 +45,8 @@ object ExtSimSetup {
     *   The actor context of this actor system.
     * @param scheduler
     *   The scheduler of simona.
+    * @param startTime
+    *   The start time of the simulation.
     * @return
     *   An [[ExtSimSetupData]] that holds information regarding the external
     *   data connections as well as the actor references of the created
@@ -55,6 +58,7 @@ object ExtSimSetup {
   )(using
       context: ActorContext[?],
       scheduler: ActorRef[SchedulerMessage],
+      startTime: ZonedDateTime,
   ): ExtSimSetupData = extLinks.zipWithIndex.foldLeft(ExtSimSetupData.apply) {
     case (extSimSetupData, (extLink, index)) =>
       // external simulation always needs at least an ExtSimAdapter
@@ -79,14 +83,14 @@ object ExtSimSetup {
         )
 
         // setup data services that belong to this external simulation
-        val updatedSetupData = connect(extSimulation, extSimSetupData)
+        val updatedSetupData = connect(extSimulation, extSimSetupData, index)
 
         // starting external simulation
         new Thread(extSimulation, s"External simulation $index")
           .start()
 
         // updating the data with newly connected external simulation
-        updatedSetupData.update(extSimAdapter)
+        updatedSetupData.updateAdapter(extSimAdapter)
       } match {
         case Failure(exception) =>
           log.warn(
@@ -104,6 +108,8 @@ object ExtSimSetup {
     *   To connect.
     * @param extSimSetupData
     *   That contains information about all external simulations.
+    * @param index
+    *   Index of the external link interface.
     * @param context
     *   The actor context of this actor system.
     * @param scheduler
@@ -116,10 +122,12 @@ object ExtSimSetup {
   private[setup] def connect(
       extSimulation: ExtSimulation,
       extSimSetupData: ExtSimSetupData,
+      index: Int,
   )(using
       context: ActorContext[?],
       scheduler: ActorRef[SchedulerMessage],
       extSimAdapterData: ExtSimAdapterData,
+      startTime: ZonedDateTime,
   ): ExtSimSetupData = {
     given extSimAdapter: ActorRef[ControlResponseMessageFromExt] =
       extSimAdapterData.getAdapter
@@ -134,8 +142,35 @@ object ExtSimSetup {
     val updatedSetupData = connections.foldLeft(extSimSetupData) {
       case (setupData, connection) =>
         connection match {
+          case extEmDataConnection: ExtEmDataConnection =>
+            if setupData.emDataService.nonEmpty then {
+              throw ServiceException(
+                s"Trying to connect another EmDataConnection. Currently only one is allowed."
+              )
+            }
+
+            if extEmDataConnection.getControlledEms.isEmpty then {
+              log.warn(
+                s"External em connection $extEmDataConnection is not used, because there are no controlled ems present!"
+              )
+              setupData
+            } else {
+              val serviceRef = context.spawn(
+                ExtEmDataService(scheduler),
+                "ExtEmDataService",
+              )
+
+              setupService(
+                extEmDataConnection,
+                serviceRef,
+                InitExtEmData(_, startTime),
+              )
+
+              extSimSetupData.update(extEmDataConnection, serviceRef)
+            }
+
           case extEvDataConnection: ExtEvDataConnection =>
-            if setupData.evDataConnection.nonEmpty then {
+            if setupData.evDataService.nonEmpty then {
               throw ServiceException(
                 s"Trying to connect another EvDataConnection. Currently only one is allowed."
               )