+++ src/main/scala/edu/ie3/simona/sim/setup/SimonaSetup.scala
@@ -38,7 +38,7 @@ trait SimonaSetup {
 
   /** Main arguments of the executable. May be used to pass additional
     * configuration parameters to the setup e.g. for external simulation
-    * configuration
+    * configuration.
     */
   val args: Array[String]
 
@@ -46,23 +46,23 @@ trait SimonaSetup {
     */
   def logOutputDir: Path
 
-  /** Creates the runtime event listener
+  /** Creates the runtime event listener.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @return
-    *   An actor reference to the runtime event listener
+    *   An actor reference to the runtime event listener.
     */
   def runtimeEventListener(
       context: ActorContext[?]
   ): ActorRef[RuntimeEventListener.Request]
 
-  /** Creates a sequence of result event listeners
+  /** Creates a sequence of result event listeners.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @return
-    *   A sequence of actor references to result event listeners
+    *   A sequence of actor references to result event listeners.
     */
   def resultEventListener(
       context: ActorContext[?]
@@ -73,14 +73,14 @@ trait SimonaSetup {
     * it's subordinate workers.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @param scheduler
-    *   Actor reference to it's according scheduler to use
+    *   Actor reference to it's according scheduler to use.
     * @param extSimSetupData
-    *   that can contain external
+    *   that can contain external.
     *   [[edu.ie3.simona.api.data.primarydata.ExtPrimaryDataConnection]]
     * @return
-    *   An actor reference to the service
+    *   An actor reference to the service.
     */
   def primaryServiceProxy(
       context: ActorContext[?],
@@ -88,46 +88,47 @@ trait SimonaSetup {
       extSimSetupData: ExtSimSetupData,
   ): ActorRef[ServiceMessage]
 
-  /** Creates a weather service
+  /** Creates a weather service.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @param scheduler
-    *   Actor reference to it's according scheduler to use
+    *   Actor reference to it's according scheduler to use.
     * @return
     *   An actor reference to the service as well as matching data to initialize
-    *   the service
+    *   the service.
     */
   def weatherService(
       context: ActorContext[?],
       scheduler: ActorRef[SchedulerMessage],
   ): ActorRef[ServiceMessage]
 
-  /** Creates a load profile service
+  /** Creates a load profile service.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @param scheduler
-    *   Actor reference to it's according scheduler to use
+    *   Actor reference to it's according scheduler to use.
     * @return
     *   An actor reference to the service as well as matching data to initialize
-    *   the service
+    *   the service.
     */
   def loadProfileService(
       context: ActorContext[?],
       scheduler: ActorRef[SchedulerMessage],
   ): ActorRef[ServiceMessage]
 
-  /** Loads external simulations and provides corresponding actors and init data
+  /** Loads external simulations and provides corresponding actors and init
+    * data.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @param scheduler
-    *   Actor reference to the scheduler to use
+    *   Actor reference to the scheduler to use.
     * @param extSimPath
-    *   option for a directory with external simulations
+    *   Option for a directory with external simulations.
     * @return
-    *   External simulations and their init data
+    *   External simulations and their init data.
     */
   def extSimulations(
       context: ActorContext[?],
@@ -135,16 +136,16 @@ trait SimonaSetup {
       extSimPath: Option[Path],
   ): ExtSimSetupData
 
-  /** Creates the time advancer
+  /** Creates the time advancer.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @param simulation
-    *   The simulation root actor ([[edu.ie3.simona.sim.SimonaSim]])
+    *   The simulation root actor ([[edu.ie3.simona.sim.SimonaSim]]).
     * @param runtimeEventListener
-    *   Runtime event listener
+    *   Runtime event listener.
     * @return
-    *   An actor reference to the time advancer
+    *   An actor reference to the time advancer.
     */
   def timeAdvancer(
       context: ActorContext[?],
@@ -152,15 +153,15 @@ trait SimonaSetup {
       runtimeEventListener: ActorRef[RuntimeEvent],
   ): ActorRef[TimeAdvancer.Request]
 
-  /** Creates a scheduler service
+  /** Creates a scheduler service.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @param parent
-    *   The parent scheduler, which could be a time advancer
+    *   The parent scheduler, which could be a time advancer.
     * @param coreFactory
     *   The factory creating a scheduler core that determines the scheduler's
-    *   behavior, defaulting to a regular scheduler
+    *   behavior, defaulting to a regular scheduler.
     * @return
     *   An actor reference to the scheduler
     */
@@ -170,17 +171,17 @@ trait SimonaSetup {
       coreFactory: CoreFactory = RegularSchedulerCore,
   ): ActorRef[SchedulerMessage]
 
-  /** Creates all the needed grid agents
+  /** Creates all the needed grid agents.
     *
     * @param context
-    *   Actor context to use
+    *   Actor context to use.
     * @param environmentRefs
     *   EnvironmentRefs to use
     * @param resultEventListeners
-    *   Listeners that await events from system participants
+    *   Listeners that await events from system participants.
     * @return
     *   A mapping from actor reference to it's according initialization data to
-    *   be used when setting up the agents
+    *   be used when setting up the agents.
     */
   def gridAgents(
       context: ActorContext[?],
@@ -190,7 +191,7 @@ trait SimonaSetup {
 
   /** SIMONA links sub grids connected by a three winding transformer a bit
     * different. Therefore, the internal node has to be set as superior node.
-    * All other gates are left unchanged
+    * All other gates are left unchanged.
     */
   protected val modifySubGridGateForThreeWindingSupport
       : SubGridGate => SubGridGate =