+++ src/main/scala/edu/ie3/simona/sim/setup/SimonaStandaloneSetup.scala
@@ -219,6 +219,7 @@ class SimonaStandaloneSetup(
     setupExtSim(extLinks, args)(using
       context,
       scheduler,
+      simonaConfig.simona.time.simStartTime,
     )
   }