+++ src/test/scala/edu/ie3/simona/agent/grid/DBFSAlgorithmSupGridSpec.scala
@@ -67,6 +67,7 @@ class DBFSAlgorithmSupGridSpec
     primaryServiceProxy = primaryService.ref,
     weather = weatherService.ref,
     loadProfiles = loadProfileService.ref,
+    emDataService = None,
     evDataService = None,
   )